+++ e2e/integration/apps/electron-builder-vite/vite.main.config.ts
@@ -1,7 +1,9 @@
 import { builtinModules } from 'node:module';
 import { defineConfig } from 'vite';
+import { datadogVitePlugin } from '@datadog/electron-sdk/vite-plugin';
 
 export default defineConfig({
+  plugins: [datadogVitePlugin()],
   build: {
     outDir: 'dist',
     emptyOutDir: false,