+++ src/assembly/commonContext.ts
@@ -24,4 +24,10 @@ export function registerCommonContext(configuration: Configuration, hooks: Forma
     application: { id: configuration.applicationId },
     _dd: { format_version: 2 },
   }));
+
+  hooks.registerSpan(() => ({
+    meta: {
+      '_dd.application.id': configuration.applicationId,
+    },
+  }));
 }